+++ R/tool_model.extract.R
@@ -54,8 +54,6 @@
 #' pGrunfeld <- pdata.frame(Grunfeld)
 #' 
 #' # then make a model frame from a formula and a pdata.frame
-##pform <- pFormula(inv ~ value + capital)
-##mf <- model.frame(pform, data = pGrunfeld)
 #' form <- inv ~ value
 #' mf <- model.frame(pGrunfeld, form)
 #' 
@@ -211,7 +209,8 @@ model.matrix.pdata.frame <- function(object,
 #' 
 #' The model response is extracted from a `pdata.frame` (where the
 #' response must reside in the first column; this is the case for a
-#' model frame), a `pFormula` + `data` or a `plm` object, and the
+#' model frame), a `Formula` + `data` (being a model frame) or a `plm` 
+#' object, and the
 #' transformation specified by `effect` and `model` is applied to
 #' it.\cr Constructing the model frame first ensures proper `NA`
 #' handling and the response being placed in the first column, see
@@ -239,7 +238,7 @@ model.matrix.pdata.frame <- function(object,
 #' data("Grunfeld", package = "plm")
 #' pGrunfeld <- pdata.frame(Grunfeld)
 #' 
-#' # then make a model frame from a pFormula and a pdata.frame
+#' # then make a model frame from a Formula and a pdata.frame
 #' form <- inv ~ value + capital
 #' mf <- model.frame(pGrunfeld, form)
 #'